finished refactor
