+++ src/wrappers.rs
@@ -19,5 +19,6 @@ pub mod xlib;
 #[cfg(all(target_os = "linux", feature = "opengl"))]
 pub mod glx;
 
+/// Wrappers and utilities around the Win32 API
 #[cfg(target_os = "windows")]
 pub mod win32;
