+++ src/wrappers/win32.rs
@@ -1,5 +1,4 @@
 pub mod h_instance;
-pub mod h_window;
-pub mod h_wnd;
 pub mod uuid;
-pub mod window_class;
+
+pub mod window;
