+++ src/wrappers/win32/window.rs
@@ -1,0 +1,57 @@
+mod data;
+mod handle;
+mod proc;
+mod window_class;
+
+use data::WindowUserData;
+pub use handle::HWnd;
+pub use proc::wnd_proc;
+use std::ptr::null_mut;
+use std::rc::Rc;
+use window_class::RegisteredClass;
+use windows_core::{Error, HSTRING};
+
+use crate::wrappers::win32::h_instance::HInstance;
+use windows_sys::Win32::Foundation::{HWND, LPARAM, LRESULT, WPARAM};
+use windows_sys::Win32::UI::WindowsAndMessaging::{CreateWindowExW, WINDOW_STYLE};
+
+pub trait WindowImpl: 'static {
+    fn after_create(&self, window: HWnd);
+    fn handle_message(
+        &self, window: HWnd, message_code: u32, w_param: WPARAM, l_param: LPARAM,
+    ) -> Option<LRESULT>;
+    fn destroy_started(&self, window: HWnd);
+}
+
+pub fn create_window<W: WindowImpl>(
+    title: &HSTRING, flags: WINDOW_STYLE, nc_width: i32, nc_height: i32, parent: HWND,
+    initializer: impl FnOnce(HWnd) -> W + 'static,
+) -> Result<HWND, Error> {
+    let instance = HInstance::get();
+    let window_class = RegisteredClass::register_new(instance, Some(wnd_proc::<W>))?;
+
+    let data = WindowUserData::new(initializer, window_class.clone());
+
+    let hwnd = unsafe {
+        CreateWindowExW(
+            0,
+            window_class.as_atom_ptr(),
+            title.as_ptr(),
+            flags,
+            0,
+            0,
+            nc_width,
+            nc_height,
+            parent,
+            null_mut(),
+            instance.as_raw(),
+            Rc::into_raw(data).cast(),
+        )
+    };
+
+    if hwnd.is_null() {
+        return Err(Error::from_win32());
+    }
+
+    Ok(hwnd)
+}
